requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
httpx==0.26.0
asyncpraw==7.7.1
pydantic==2.6.1
pydantic-settings==2.1.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1 